+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Make required fields not required
+
 ### 1.0.1: August 27th, 2018
 * Remove sensitive debug info
 * Avoid calls if project ID not set